requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.0
# Browser automation
playwright==1.58.0
# Apify SDK for actor compatibility
apify==1.6.0